requirements.txt
APScheduler==3.10.4
Authlib==1.6.0
blinker==1.8.2
cachetools==6.0.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.7
contourpy==1.3.0
cryptography==45.0.3
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Faker==37.4.0
Flask==3.0.3
Flask-Mail==0.10.0
flask-paginate==2024.4.12
fonttools==4.54.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
lxml==5.4.0
MarkupSafe==2.1.5
matplotlib==3.9.2
more-itertools==10.7.0
mysql-connector==2.2.9
mysql-connector-python==8.4.0
numpy==2.2.6
packaging==24.1
pdfkit==1.0.0
pillow==10.4.0
plotly==5.24.1
premailer==3.10.0
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
requests==2.32.3
secure-smtplib==0.1.1
six==1.16.0
tenacity==9.1.2
tzdata==2025.2
tzlocal==5.2
urllib3==2.4.0
Werkzeug==3.0.3
